+++ src/products/components/ProductListPage/filters.ts
@@ -1,5 +1,4 @@
 import { IFilter } from "@saleor/components/Filter";
-import { MultiAutocompleteChoiceType } from "@saleor/components/MultiAutocompleteSelectField";
 import { commonMessages, sectionNames } from "@saleor/intl";
 import { AutocompleteFilterOpts, FilterOpts, MinMax } from "@saleor/types";
 import {
@@ -26,12 +25,13 @@ export enum ProductFilterKeys {
 export interface ProductListFilterOpts {
   attributes: Array<
     FilterOpts<string[]> & {
-      choices: MultiAutocompleteChoiceType[];
+      id: string;
       name: string;
       slug: string;
       inputType: AttributeInputTypeEnum;
     }
   >;
+  attributeChoices: FilterOpts<string[]> & AutocompleteFilterOpts;
   categories: FilterOpts<string[]> & AutocompleteFilterOpts;
   collections: FilterOpts<string[]> & AutocompleteFilterOpts;
   price: FilterOpts<MinMax>;
@@ -178,12 +178,21 @@ export function createFilterStructure(
       group: ProductFilterKeys.attributes
     })),
     ...defaultAttributes.map(attr => ({
-      ...createOptionsField(
+      ...createAutocompleteField(
         attr.slug as any,
         attr.name,
         attr.value,
+        opts.attributeChoices.displayValues,
         true,
-        attr.choices
+        opts.attributeChoices.choices,
+        {
+          hasMore: opts.attributeChoices.hasMore,
+          initialSearch: "",
+          loading: opts.attributeChoices.loading,
+          onFetchMore: opts.attributeChoices.onFetchMore,
+          onSearchChange: opts.attributeChoices.onSearchChange
+        },
+        attr.id
       ),
       active: attr.active,
       group: ProductFilterKeys.attributes